Improved Data Retrieval + /restart command
The number of times that data is retrieved from our database is reduced considerably, which speeds up the "Manage websites" request by a significant amount fo time. In addition, a /restart command is added to allow users to reset the bot in the event where they are stuck somewhere at a point where there is an unexpected bug.

Following that, the manage websites option now presents a condensed list of options to users (manage posts/listings/agent details) which makes the whole prompt/response flow much neater.